// app/protogen/pebble.d.ts
export interface ISensorData {

  /** SensorData snr */
  snr?: (number | null);

  /** SensorData vbat */
  vbat?: (number | null);

  /** SensorData latitude */
  latitude?: (number | null);

  /** SensorData longitude */
  longitude?: (number | null);

  /** SensorData gasResistance */
  gasResistance?: (number | null);

  /** SensorData temperature */
  temperature?: (number | null);

  /** SensorData pressure */
  pressure?: (number | null);

  /** SensorData humidity */
  humidity?: (number | null);

  /** SensorData light */
  light?: (number | null);

  /** SensorData temperature2 */
  temperature2?: (number | null);

  /** SensorData gyroscope */
  gyroscope?: (number[] | null);

  /** SensorData accelerometer */
  accelerometer?: (number[] | null);

  /** SensorData random */
  random?: (string | null);
}

/** Represents a SensorData. */
export class SensorData implements ISensorData {

  /**
   * Constructs a new SensorData.
   * @param [properties] Properties to set
   */
  constructor(properties?: ISensorData);

  /** SensorData snr. */
  public snr: number;

  /** SensorData vbat. */
  public vbat: number;

  /** SensorData latitude. */
  public latitude: number;

  /** SensorData longitude. */
  public longitude: number;

  /** SensorData gasResistance. */
  public gasResistance: number;

  /** SensorData temperature. */
  public temperature: number;

  /** SensorData pressure. */
  public pressure: number;

  /** SensorData humidity. */
  public humidity: number;

  /** SensorData light. */
  public light: number;

  /** SensorData temperature2. */
  public temperature2: number;

  /** SensorData gyroscope. */
  public gyroscope: number[];

  /** SensorData accelerometer. */
  public accelerometer: number[];

  /** SensorData random. */
  public random: string;

  /**
   * Creates a new SensorData instance using the specified properties.
   * @param [properties] Properties to set
   * @returns SensorData instance
   */
  public static create(properties?: ISensorData): SensorData;

  /**
   * Encodes the specified SensorData message. Does not implicitly {@link SensorData.verify|verify} messages.
   * @param message SensorData message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encode(message: ISensorData, writer?: $protobuf.Writer): $protobuf.Writer;

  /**
   * Encodes the specified SensorData message, length delimited. Does not implicitly {@link SensorData.verify|verify} messages.
   * @param message SensorData message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encodeDelimited(message: ISensorData, writer?: $protobuf.Writer): $protobuf.Writer;

  /**
   * Decodes a SensorData message from the specified reader or buffer.
   * @param reader Reader or buffer to decode from
   * @param [length] Message length if known beforehand
   * @returns SensorData
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decode(reader: ($protobuf.Reader | Uint8Array), length?: number): SensorData;

  /**
   * Decodes a SensorData message from the specified reader or buffer, length delimited.
   * @param reader Reader or buffer to decode from
   * @returns SensorData
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decodeDelimited(reader: ($protobuf.Reader | Uint8Array)): SensorData;

  /**
   * Verifies a SensorData message.
   * @param message Plain object to verify
   * @returns `null` if valid, otherwise the reason why it is not
   */
  public static verify(message: { [k: string]: any }): (string | null);

  /**
   * Creates a SensorData message from a plain object. Also converts values to their respective internal types.
   * @param object Plain object
   * @returns SensorData
   */
  public static fromObject(object: { [k: string]: any }): SensorData;

  /**
   * Creates a plain object from a SensorData message. Also converts values to other types if specified.
   * @param message SensorData
   * @param [options] Conversion options
   * @returns Plain object
   */
  public static toObject(message: SensorData, options?: $protobuf.IConversionOptions): { [k: string]: any };

  /**
   * Converts this SensorData to JSON.
   * @returns JSON object
   */
  public toJSON(): { [k: string]: any };
}

/** Properties of a SensorConfig. */
export interface ISensorConfig {

  /** SensorConfig bulkUpload */
  bulkUpload?: (number | null);

  /** SensorConfig dataChannel */
  dataChannel?: (number | null);

  /** SensorConfig uploadPeriod */
  uploadPeriod?: (number | null);

  /** SensorConfig bulkUploadSamplingCnt */
  bulkUploadSamplingCnt?: (number | null);

  /** SensorConfig bulkUploadSamplingFreq */
  bulkUploadSamplingFreq?: (number | null);

  /** SensorConfig beep */
  beep?: (number | null);

  /** SensorConfig firmware */
  firmware?: (string | null);
}

/** Represents a SensorConfig. */
export class SensorConfig implements ISensorConfig {

  /**
   * Constructs a new SensorConfig.
   * @param [properties] Properties to set
   */
  constructor(properties?: ISensorConfig);

  /** SensorConfig bulkUpload. */
  public bulkUpload: number;

  /** SensorConfig dataChannel. */
  public dataChannel: number;

  /** SensorConfig uploadPeriod. */
  public uploadPeriod: number;

  /** SensorConfig bulkUploadSamplingCnt. */
  public bulkUploadSamplingCnt: number;

  /** SensorConfig bulkUploadSamplingFreq. */
  public bulkUploadSamplingFreq: number;

  /** SensorConfig beep. */
  public beep: number;

  /** SensorConfig firmware. */
  public firmware: string;

  /**
   * Creates a new SensorConfig instance using the specified properties.
   * @param [properties] Properties to set
   * @returns SensorConfig instance
   */
  public static create(properties?: ISensorConfig): SensorConfig;

  /**
   * Encodes the specified SensorConfig message. Does not implicitly {@link SensorConfig.verify|verify} messages.
   * @param message SensorConfig message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encode(message: ISensorConfig, writer?: $protobuf.Writer): $protobuf.Writer;

  /**
   * Encodes the specified SensorConfig message, length delimited. Does not implicitly {@link SensorConfig.verify|verify} messages.
   * @param message SensorConfig message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encodeDelimited(message: ISensorConfig, writer?: $protobuf.Writer): $protobuf.Writer;

  /**
   * Decodes a SensorConfig message from the specified reader or buffer.
   * @param reader Reader or buffer to decode from
   * @param [length] Message length if known beforehand
   * @returns SensorConfig
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decode(reader: ($protobuf.Reader | Uint8Array), length?: number): SensorConfig;

  /**
   * Decodes a SensorConfig message from the specified reader or buffer, length delimited.
   * @param reader Reader or buffer to decode from
   * @returns SensorConfig
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decodeDelimited(reader: ($protobuf.Reader | Uint8Array)): SensorConfig;

  /**
   * Verifies a SensorConfig message.
   * @param message Plain object to verify
   * @returns `null` if valid, otherwise the reason why it is not
   */
  public static verify(message: { [k: string]: any }): (string | null);

  /**
   * Creates a SensorConfig message from a plain object. Also converts values to their respective internal types.
   * @param object Plain object
   * @returns SensorConfig
   */
  public static fromObject(object: { [k: string]: any }): SensorConfig;

  /**
   * Creates a plain object from a SensorConfig message. Also converts values to other types if specified.
   * @param message SensorConfig
   * @param [options] Conversion options
   * @returns Plain object
   */
  public static toObject(message: SensorConfig, options?: $protobuf.IConversionOptions): { [k: string]: any };

  /**
   * Converts this SensorConfig to JSON.
   * @returns JSON object
   */
  public toJSON(): { [k: string]: any };
}

/** Properties of a SensorState. */
export interface ISensorState {

  /** SensorState state */
  state?: (number | null);
}

/** Represents a SensorState. */
export class SensorState implements ISensorState {

  /**
   * Constructs a new SensorState.
   * @param [properties] Properties to set
   */
  constructor(properties?: ISensorState);

  /** SensorState state. */
  public state: number;

  /**
   * Creates a new SensorState instance using the specified properties.
   * @param [properties] Properties to set
   * @returns SensorState instance
   */
  public static create(properties?: ISensorState): SensorState;

  /**
   * Encodes the specified SensorState message. Does not implicitly {@link SensorState.verify|verify} messages.
   * @param message SensorState message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encode(message: ISensorState, writer?: $protobuf.Writer): $protobuf.Writer;

  /**
   * Encodes the specified SensorState message, length delimited. Does not implicitly {@link SensorState.verify|verify} messages.
   * @param message SensorState message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encodeDelimited(message: ISensorState, writer?: $protobuf.Writer): $protobuf.Writer;

  /**
   * Decodes a SensorState message from the specified reader or buffer.
   * @param reader Reader or buffer to decode from
   * @param [length] Message length if known beforehand
   * @returns SensorState
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decode(reader: ($protobuf.Reader | Uint8Array), length?: number): SensorState;

  /**
   * Decodes a SensorState message from the specified reader or buffer, length delimited.
   * @param reader Reader or buffer to decode from
   * @returns SensorState
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decodeDelimited(reader: ($protobuf.Reader | Uint8Array)): SensorState;

  /**
   * Verifies a SensorState message.
   * @param message Plain object to verify
   * @returns `null` if valid, otherwise the reason why it is not
   */
  public static verify(message: { [k: string]: any }): (string | null);

  /**
   * Creates a SensorState message from a plain object. Also converts values to their respective internal types.
   * @param object Plain object
   * @returns SensorState
   */
  public static fromObject(object: { [k: string]: any }): SensorState;

  /**
   * Creates a plain object from a SensorState message. Also converts values to other types if specified.
   * @param message SensorState
   * @param [options] Conversion options
   * @returns Plain object
   */
  public static toObject(message: SensorState, options?: $protobuf.IConversionOptions): { [k: string]: any };

  /**
   * Converts this SensorState to JSON.
   * @returns JSON object
   */
  public toJSON(): { [k: string]: any };
}
// ...
export namespace BinPackage {

  /** PackageType enum. */
  enum PackageType {
    DATA = 0,
    CONFIG = 1,
    STATE = 2
  }
}